accounts-db: Move cache eviction test to integration tests
`test_read_only_accounts_cache_eviction` takes long time to execute
(from 20s to over 60s, depending on the hardware), but there is no way
to make it faster without compromising on its accuracy and making sure
that the error margin in cache eviction stays low.

To make life of people running `cargo test` frequently a bit easier,
move it out to `tests/` directory, so it can be omitted when running
`cargo test --lib`.
